refactor: move bmt and span to chunk folder
